[DAD-1137] feat: 유저 정보 테이블 분리
