+++ src/pages/UserPage.tsx
@@ -9,6 +9,7 @@ import { useGetUserInformation } from '@/api/user';
 import { useConvertUnixTimeToDateFormat, useGetDaysDiff } from '@/hooks/useCalculateDateDiff';
 
 import { DarkWaveVector, EditPencilSquareIcon, LightWaveVector } from '@/components/atoms/Icon';
+import UserInfoTable from '@/components/molcules/UserPage/UserInfoTable';
 
 function UserPage() {
     const { userInformation, isGetUserInformationLoading } = useGetUserInformation();
@@ -21,72 +22,6 @@ function UserPage() {
     }
 
     const { profileUrl, name, provider, nickname, createdAt } = userInformation;
-
-    const createdDateInDateFormat = useConvertUnixTimeToDateFormat(createdAt);
-    const daysDiffFromCreatedDate = useGetDaysDiff(createdAt);
-    const daysDiffInfoString = `(+ 가입한 지 ${daysDiffFromCreatedDate}일이 지났습니다.)`;
-    const providerInfoString = `${provider} 계정으로 가입되셨습니다.`;
-
-    function userPageMenuNameAndContent(
-        name: string,
-        content: string | number | string[]
-    ) {
-        return { name, content };
-    }
-
-    function CreatedDateInfo() {
-        return <Box
-            sx={{
-                display: 'flex',
-                alignItems: {
-                    xs: 'flex-start',
-                    sm: 'center',
-                },
-                flexDirection: {
-                    xs: 'column',
-                    sm: 'row',
-                },
-                gap: {
-                    xs: '0',
-                    sm: '4px',
-                }
-            }}
-        >
-            <Typography
-                sx={{
-                    ...userPageMenuTypographyStyle,
-                    fontWeight: '400',
-                }}
-            >
-                {createdDateInDateFormat}
-            </Typography>
-            <Typography
-                sx={countDiffFromCreatedDateTypographyStyle}
-            >
-                {daysDiffInfoString}
-            </Typography>
-        </Box>
-    }
-
-    const userPageMenus = [
-        userPageMenuNameAndContent('이름', name),
-        userPageMenuNameAndContent('닉네임', nickname),
-        userPageMenuNameAndContent('가입날짜', [createdDateInDateFormat, daysDiffInfoString]),
-        userPageMenuNameAndContent('연결된 소셜 계정', providerInfoString),
-    ];
-
-    const userPageMenuTypographyStyle = {
-        color: theme.color.Gray_090,
-        fontSize: '16px',
-        lineHeight: '150%',
-    };
-
-    const countDiffFromCreatedDateTypographyStyle = {
-        color: theme.color.Gray_080,
-        fontSize: '14px',
-        fontWeight: '400',
-        lineHeight: '150%',
-    };
 
     return (
         <Wrapper>
@@ -133,48 +68,7 @@ function UserPage() {
                             boxSizing: 'border-box',
                         }}
                     >
-                        <TableContainer component={Box}
-                            sx={{
-                                width: '100%',
-                                backgroundColor: 'transparent',
-                                border: 'none',
-                            }}
-                        >
-                            <Table>
-                                <TableBody>
-                                    {userPageMenus.map((menu) => (
-                                        <TableRow>
-                                            <TableCell
-                                                component="th"
-                                                scope="row"
-                                                sx={{
-                                                    ...userPageMenuTypographyStyle,
-                                                    fontWeight: '600',
-                                                    border: 'none',
-                                                    p: '7px 20px 7px 0',
-                                                    width: 'max-content',
-                                                }}
-                                            >
-                                                {menu.name}
-                                            </TableCell>
-                                            <TableCell
-                                                sx={{
-                                                    ...userPageMenuTypographyStyle,
-                                                    fontWeight: '400',
-                                                    border: 'none',
-                                                    p: '7px 0px',
-                                                }}
-                                            >
-                                                {typeof (menu.content) === 'string'
-                                                    ? menu.content
-                                                    : <CreatedDateInfo />
-                                                }
-                                            </TableCell>
-                                        </TableRow>
-                                    ))}
-                                </TableBody>
-                            </Table>
-                        </TableContainer>
+                        <UserInfoTable userInformation={userInformation} />
                     </Box>
                     <Button
                         variant='outlined'
